+++ resolwe/flow/tasks.py
@@ -4,5 +4,5 @@ from celery import shared_task
 
 
 @shared_task
-def celery_run(executor, data_id, script):
-    executor.run(data_id, script)
+def celery_run(executor, data_id, script, verbosity):
+    executor.run(data_id, script, verbosity)