+++ docs/CHANGELOG.rst
@@ -17,6 +17,7 @@ Fixed
 -----
 - All Resolwe's environment variables are passed to Tox's testing environment
 - Included all source files and supplementary package data in sdist
+- Made Celery engine work
 
 
 ==================